app/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
fastapi==0.127.0
greenlet==3.3.0
h11==0.16.0
idna==3.11
mysql-connector-python==9.5.0
pydantic==2.12.5
pydantic_core==2.41.5
PyMySQL==1.1.2
requests==2.32.5
SQLAlchemy==2.0.45
sqlmodel==0.0.27
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.2
uvicorn==0.40.0
